Add files via upload
This .ZIP folder has everything you need to create your own USB Stealer. Inside, you will see some .vbs scripts, 2 batch files, and some folders. 

.VBS Files- 
These are used to silently execute their attached .bat files. Use these to launch the stealers. 

av.bat and all.bat- 
av.bat is a batch file THAT MUST BE RAN WITH ADMINISTRATOR PRIVILEGES that will set windows defender exclusions for D: drive and C: drive. Use when applicable. You may change the D: to whatever you call your USB drive name is. All.bat just runs all the .vbs at the same time for ease of use, besides the files one. I didn't put the file stealer in the batch due to me not valuing stealing files as a priority. If you desire to do so, add that to the code of the batch file. 

The folders- 
These folders are where all the magic happens. First, let's start with the wifiSTEALER. This program's goal is to steal all the cached passwords in said computer. Inside the folder is an exe, py, and 2 batch files. The killwifi batch code was in response to an error where the exe file would stay open after the code was finished running. The .exe itself just steals the wifi passwords, this takes a couple seconds. The py file is the actual code of the .exe, and was included in case you do not trust me with the exe file. Just use a converter to change the py to exe if you want to do that. The second batch file just runs the exe and saves the stolen passwords to the PASS folder. 

The RAT folder is where you may put any additional exe's that you may want to execute. I specifically may want to execute RAT files on the victim so I called the folder RAT. Anyways, inside of the folder are two batch files and a text file. Ratlaunch is what starts your RAT or other exe files in the folder. The TokenStealer.bat steals someone's discord token and sends it to a webhook of your desire. The text file just clarifies some instructions. 

Next, is the Product Key folder. There is nothing in the folder, as the vbs file is all you need. It just steals the product key from the computer and saves it into the folder. It will prompt up a message box, and you can just dismiss it. 

The passwordSTEALER is something that steals many saved passwords in a computer. These exes were from Nirsoft and steal a variety of passwords, including browser passes and even browser history. the batch file just executes all the exes at once. The passwords are saved to text files in the PASS folder. 

The IpStealer takes the IP of the victim's connected wifi and saves it into the ips folder. 

Finally, the fileSTEALER folder takes any desired files from the victim. You may change the code to steal whatever files you want, and they will be saved into the FIles folder. 

Well, that's it for the stealer. Any questions or suggestions, feel free to contact me at my discord, SOVIET#3565. Have fun bois! :)